dw-uploader/requirements.txt
IntelHex
pyserial
progressbar2
pyinstaller
